+++ src/gemseo_pymoo/post/base_weighted_pymoo_post_algorithms_settings.py
@@ -17,7 +17,6 @@
 from __future__ import annotations
 
 from functools import partial
-from typing import Any
 
 from gemseo.post.base_post_settings import BasePostSettings
 from gemseo.utils.pydantic import copy_field
@@ -41,17 +40,18 @@ class WeightedPostSettings(BasePymooPostSettings):
     classes.
     """
 
-    decomposition: Any | None = Field(
+    decomposition: Decomposition | None = Field(
         default=None,
         description="The instance of the scalarization function to use. "
-        "If ``None``, use a weighted sum.",
+        "If `None`, use a weighted sum.",
     )
 
-    weights: NDArrayPydantic | NDArrayPydantic[NDArrayPydantic] | None = Field(
+    weights: NDArrayPydantic[float] | None = Field(
         default=None,
-        description="The weights for the scalarization function. If None, a "
-        "normalized array is used, e.g. [1./n, 1./n, ..., 1./n] "
-        "for an optimization problem with n-objectives.",
+        description="The weights for the scalarization function. If `None`, a "
+        "normalized array is used, "
+        "e.g. $[1./n, 1./n, ..., 1./n]$ "
+        "for an optimization problem with $n$ objectives.",
     )
 
     # scalar opts
@@ -71,19 +71,22 @@ class WeightedPostSettings(BasePymooPostSettings):
 
     @field_validator("weights")
     @classmethod
-    def __check_points(
-        cls, weights: NDArrayPydantic | NDArrayPydantic[NDArrayPydantic] | None
-    ):
+    def __check_weights(
+        cls, weights: NDArrayPydantic[float] | None
+    ) -> NDArrayPydantic[float] | None:
         """Check the size of the weights setting arrays."""
         return _array_validation_function(weights)
 
     @field_validator("decomposition", mode="before")
     @classmethod
-    def __check_decomposition(cls, decomposition: Decomposition | None):
+    def __check_decomposition(
+        cls, decomposition: Decomposition | None
+    ) -> Decomposition:
         """Check for the type of the decomposition setting."""
         if decomposition is None:
-            decomposition = WeightedSum()
-        elif not isinstance(decomposition, Decomposition):
+            return WeightedSum()
+
+        if not isinstance(decomposition, Decomposition):
             msg = (
                 "The scalarization function must be an instance of "
                 "pymoo.core.Decomposition."